% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_data_check.R
\name{initial_data_check}
\alias{initial_data_check}
\title{Find NAs and duplicates in dataframe}
\usage{
initial_data_check(df)
}
\arguments{
\item{df}{A data frame to inspect}
}
\value{
A nested list with number and indices of NAs and duplicates
}
\description{
Find NAs and duplicates in dataframe
}
\examples{
initial_data_check(df)
}
